requirements.txt
beautifulsoup4==4.9.3
requests==2.25.1
notion-client==1.0.0
python-dotenv==0.19.2